+++ src/components/Movies/Movies.jsx
@@ -5,15 +5,11 @@ import MoviesCardList from './MoviesCardList'; // jsx
 
 import MoviesApi from '../../utils/MoviesApi'; // апи к фильмам
 import NormCard from '../../utils/NormCard'; // функция для создания карточки для моего апи
+import SearchMovies from '../../utils/SearchMovies'; // поиск фильмов
 
 function Movies() {
   const [cardsBeatfilm, setCardsBeatfilm] = useState([]); // все начальные карточки
-  const [foundMovies, setFoundMovies] = useState([]); // все начальные карточки
-
-  // удаление имейла вошедшего пользователя из локального хранилища
-  useEffect(() => {
-    localStorage.removeItem('potentialUserEmail');
-  }, []);
+  const [foundMovies, setFoundMovies] = useState([]); // найденные карточки из локального хранилища, иначе пустой массив
 
   // <-- Функция загрузки всех фильмов
   function loadAllMovies() {
@@ -30,48 +26,20 @@ function Movies() {
   }
   // Функция загрузки всех фильмов -- />
 
-  // загрузим все фильмы до сабмита чтоб они были сразу доступны для фильтрации при поиске
   useEffect(() => {
-    loadAllMovies();
+    loadAllMovies(); // загрузим все фильмы до сабмита чтоб они были сразу доступны для фильтрации при поиске
+    // setFoundMovies(localStorage.getItem('moviesState').foundMovies);
+    localStorage.removeItem('potentialUserEmail'); // удаление имейла вошедшего пользователя из локального хранилища
   }, []);
 
-  // <-- Функция поиска
+  // <-- Обработчика сабмита поиска
   function handleSearch(inputData, shortChecked) {
-    let foundArray;
-
-    function filterItemsObject(el, query) {
-      let result = false;
-
-      for (let key in el) {
-        if (
-          el[key]
-            .toString()
-            .toLowerCase()
-            .indexOf(query.toString().toLowerCase()) !== -1
-        ) {
-          result = true;
-        }
-      }
-      return result;
-    }
-
-    function filterItemsArray(arr, query) {
-      return arr.filter((el) => filterItemsObject(el, query));
-    }
-
-    if (shortChecked) {
-      foundArray = cardsBeatfilm.filter((e) => e.duration <= 40);
-    } else {
-      foundArray = cardsBeatfilm;
-    }
-    const finalArray = filterItemsArray(foundArray, inputData);
-    setFoundMovies(finalArray);
-    // Функция поиска -- />
+    setFoundMovies(SearchMovies(inputData, shortChecked, cardsBeatfilm));
   }
 
   return (
     <main className="movies content">
-      <SearchForm loadAllMovies={loadAllMovies} handleSearch={handleSearch} />
+      <SearchForm handleSearch={handleSearch} foundMovies={foundMovies} />
       <MoviesCardList foundMovies={foundMovies} />
     </main>
   );
